Rework module information retrieval
Filesystem slugs were not retrieved properly, I hope to have fixed this by sourcing the filesystem slugs (plural) from the `make_flyable` function calls in the plugin files.
Other than that the module parsing is now done in a separate class, perhaps that I should make it static again but for now I'm keeping it like this.