Reworked directories (now virtual rep.), started working on liveries
Directories are now somewhat virtually representing the directories they represent, in structure.
Added listing installed liveries to the GUI, maybe eventually this can be done in the TreeView. Has to be decided on, can also do both using an option in the settings dialog. Dunno. Who knows. Not me. Not yet.