Transform Installable into the more dynamic Plugin base
Zipped liveries support was added, but after some testing this is apparently a feature only available to liveries installed in the installation directory, under CoreMods. That's a directory we don't touch yet.
For the rest, Plugin base allows use for both installed and to-be installed plugins. Meaning this will form the basis for mod/livery installation.